requirements.txt
opencv-python
numpy
tensorflow
rawpy
imageio
pywin32
pyserial